backend/requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.12.5
email-validator